v0.1.1: Fixed Windows errors and fixed operators Fixed some Windows specific code breaking GNU Linux code Fixed multiple operator expressions Added ',' to the lexer blacklist

+++ src/types.h
@@ -208,7 +208,7 @@ namespace vtex
                 if (RHS.type() != number)
                     return {};
                 
-                return LFloat(std::fmod(Val, *(long double*)RHS.get()));
+                return LFloat(fmodl(Val, *(long double*)RHS.get()));
             }
             Type greater(Type RHS) override
             {
